Update Calsim3 inputs for v0.9
This commit updates input files of Calsim3 DeltaCD for a new upcoming
version of DeltaCD v0.9. This version now uses the original 168 subareas
in DETAW, and DETAW outputs get split in the DCD part, similarly to
the previous DETAW-DCD. To achieve this, new input files are introduced,
and some file formats and contents got updated.
Note that this input files do not work with v0.8.

DETAW-96, DETAW-94