Create historical branch
We are creating branches for different settings, for example, historical
and planning inputs. This branch contains historical inputs for DSM2 and
SCHISM. Delta and Suisun Marsh inputs are separated into two
directories. Files are just moved and slightly renamed. No changes are
made in file contents.
We do not need the GitHub action for documenation in branches, so, it is
removed with docsrc directory.