Create planning branch
We are creating branches for different settings, for example, historical
and planning inputs. This branch contains planning inputs for Calsim3.
Files are just moved and slightly renamed. No changes are made in file
contents.
We do not need the GitHub action for documenation in branches, so, it is
removed with docsrc directory.